delete the old app reducer, stick with the new
