--- applications/desktop/src/notebook/reducers/app.js
@@ -1,20 +1,0 @@
-// @flow
-import { remote } from "electron";
-
-import { app } from "@nteract/core/reducers";
-
-import {
-  makeAppRecord,
-  makeDesktopHostRecord,
-  AppRecord
-} from "@nteract/types/core/records";
-
-export default function handleApp(
-  state: AppRecord = makeAppRecord({
-    host: makeDesktopHostRecord(),
-    version: remote.app.getVersion()
-  }),
-  action: *
-) {
-  return app(state, action);
-}
