+++ packages/core/__tests__/app-spec.js
@@ -59,3 +59,50 @@ describe("setExecutionState", () => {
     expect(state.kernel.status).toBe("idle");
   });
 });
+
+describe("setNotificationSystem", () => {
+  test("returns the same originalState if notificationSystem is undefined", () => {
+    const originalState = makeAppRecord({
+      kernel: makeLocalKernelRecord({
+        channels: false,
+        spawn: false,
+        connectionFile: false
+      })
+    });
+
+    const action = { type: actionTypes.SET_NOTIFICATION_SYSTEM };
+
+    const state = app(originalState, action);
+    expect(state.notificationSystem).toBeNull();
+  });
+  test("sets the notificationSystem if given", () => {
+    const originalState = makeAppRecord({
+      kernel: makeLocalKernelRecord({
+        channels: false,
+        spawn: false,
+        connectionFile: false
+      })
+    });
+
+    const action = {
+      type: actionTypes.SET_NOTIFICATION_SYSTEM,
+      notificationSystem: ""
+    };
+
+    const state = app(originalState, action);
+    expect(state.notificationSystem).toBe("");
+  });
+});
+
+describe("setGithubToken", () => {
+  test("calls setGithubToken", () => {
+    const originalState = makeAppRecord({
+      githubToken: null
+    });
+
+    const action = { type: actionTypes.SET_GITHUB_TOKEN, githubToken: "TOKEN" };
+
+    const state = app(originalState, action);
+    expect(state.githubToken).toEqual("TOKEN");
+  });
+});
